Remove all content from MainWindow.xaml and MainWindow.xaml.cs
The entire content of the MainWindow.xaml file has been removed,
including the Window element and its attributes, as well as the
Grid element inside it. Similarly, the entire content of the
MainWindow.xaml.cs file has been removed, including all the using
directives, the Cubicon namespace, and the MainWindow class with
its constructor and InitializeComponent method.